Rename helper to lib
