+++ packages/contentful-ssg/src/lib/array.ts
