+++ packages/contentful-ssg/src/lib/config.ts
