+++ packages/contentful-ssg/src/lib/contentful.ts
