+++ packages/contentful-ssg/src/lib/create-require.ts
