+++ packages/contentful-ssg/src/lib/error.ts
