+++ packages/contentful-ssg/src/lib/file-manager.ts
