+++ packages/contentful-ssg/src/lib/hook-manager.test.ts
