+++ packages/contentful-ssg/src/lib/hook-manager.ts
