+++ packages/contentful-ssg/src/lib/object.test.ts
