+++ packages/contentful-ssg/src/lib/object.ts
