+++ packages/contentful-ssg/src/lib/stats.ts
