+++ packages/contentful-ssg/src/lib/utils.ts
